Moved logic handleContractNegotiationAgreed from handler to apiService
