+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.0.1] - 2024-07-03
+## [0.0.1] - 2024-07-08
 
 ### Added
