+++ negotiation/src/main/java/it/eng/negotiation/service/ContractNegotiationAPIService.java
@@ -1,5 +1,7 @@
 package it.eng.negotiation.service;
 
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.Collection;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -12,7 +14,9 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import it.eng.negotiation.exception.ContractNegotiationAPIException;
+import it.eng.negotiation.exception.ContractNegotiationInvalidStateException;
 import it.eng.negotiation.exception.ContractNegotiationNotFoundException;
+import it.eng.negotiation.exception.OfferNotFoundException;
 import it.eng.negotiation.model.Agreement;
 import it.eng.negotiation.model.ContractAgreementMessage;
 import it.eng.negotiation.model.ContractNegotiation;
@@ -36,6 +40,8 @@ import lombok.extern.slf4j.Slf4j;
 @Service
 @Slf4j
 public class ContractNegotiationAPIService {
+
+	private static final DateTimeFormatter FORMATTER = DateTimeFormatter.ISO_OFFSET_DATE_TIME;
 
 	private final OkHttpRestClient okHttpRestClient;
 	private final ContractNegotiationRepository contractNegotiationRepository;
@@ -272,5 +278,63 @@ public class ContractNegotiationAPIService {
 			log.error("Error response received!");
 			throw new ContractNegotiationAPIException(response.getMessage());
 		}
+	}
+	
+	/**
+	 * Negotiate status to AGREED after successful reponse from connector
+	 * @param contractNegotiationId
+	 * @return
+	 */
+	public ContractNegotiation handleContractNegotiationAgreed(String contractNegotiationId) {
+		ContractNegotiation contractNegotiation = contractNegotiationRepository.findById(contractNegotiationId)
+        	.orElseThrow(() ->
+                new ContractNegotiationNotFoundException("Contract negotiation with id " + contractNegotiationId + " not found"));
+	
+		// TODO check if this is valid check
+		if (!contractNegotiation.getState().canTransitTo(ContractNegotiationState.AGREED)) {
+			throw new ContractNegotiationInvalidStateException(
+					"Contract negotiation with providerPid " + contractNegotiation.getProviderPid() + 
+					" and consumerPid " + contractNegotiation.getConsumerPid() + " is not in REQUESTED state, aborting verification", 
+					contractNegotiation.getConsumerPid(), contractNegotiation.getProviderPid());
+		}
+
+		Offer offer = contractNegotiation.getOffer();
+		offerRepository.findById(offer.getId())
+       		.orElseThrow(() ->
+       			new OfferNotFoundException("Offer with id " + offer.getId() + " binded to contract neogitation not found"));
+
+		ContractAgreementMessage agreementMessage = ContractAgreementMessage.Builder.newInstance()
+				.consumerPid(contractNegotiation.getConsumerPid())
+				.providerPid(contractNegotiation.getProviderPid())
+				.callbackAddress(properties.providerCallbackAddress())
+				.agreement(agreementFromOffer(offer, contractNegotiation.getAssigner()))
+				.build();
+		// TODO this one will fail because provider does not have consumer callbackAddress for sending agreement
+		GenericApiResponse<String> response = okHttpRestClient.sendRequestProtocol(
+				ContractNegotiationCallback.getContractAgreementCallback(contractNegotiation.getCallbackAddress(), contractNegotiation.getConsumerPid()), 
+				Serializer.serializeProtocolJsonNode(agreementMessage),
+				credentialUtils.getConnectorCredentials());
+		if(response.isSuccess()) {
+			log.info("Updating status for negotiation {} to agreed", contractNegotiation.getId());
+			ContractNegotiation contractNegtiationAgreed = contractNegotiation.withNewContractNegotiationState(ContractNegotiationState.AGREED);
+			contractNegotiationRepository.save(contractNegtiationAgreed);
+			log.info("Saving agreement..." + agreementMessage.getAgreement().getId());
+			agreementRepository.save(agreementMessage.getAgreement());
+			return contractNegtiationAgreed;
+		} else {
+			log.error("Response status not 200 - consumer did not process AgreementMessage correct");
+			throw new ContractNegotiationAPIException("consumer did not process AgreementMessage correct");
+		}
+	}
+	
+	private Agreement agreementFromOffer(Offer offer, String assigner) {
+		return Agreement.Builder.newInstance()
+				.id(UUID.randomUUID().toString())
+				.assignee(properties.getAssignee())
+				.assigner(assigner)
+				.target(offer.getTarget())
+				.timestamp(FORMATTER.format(ZonedDateTime.now()))
+				.permission(offer.getPermission())
+				.build();
 	}
 }
