+++ negotiation/src/test/java/it/eng/negotiation/service/ContractNegotiationAPIServiceTest.java
@@ -26,6 +26,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import com.fasterxml.jackson.databind.JsonNode;
 
 import it.eng.negotiation.exception.ContractNegotiationAPIException;
+import it.eng.negotiation.exception.ContractNegotiationInvalidStateException;
+import it.eng.negotiation.exception.OfferNotFoundException;
 import it.eng.negotiation.model.Agreement;
 import it.eng.negotiation.model.ContractNegotiation;
 import it.eng.negotiation.model.ContractNegotiationState;
@@ -60,6 +62,8 @@ public class ContractNegotiationAPIServiceTest {
 	
 	@Captor
 	private ArgumentCaptor<ContractNegotiation> argumentCaptor;
+	@Captor
+	private ArgumentCaptor<Agreement> argCaptorAgreement;
 
 	@InjectMocks
 	private ContractNegotiationAPIService service;
@@ -311,5 +315,65 @@ public class ContractNegotiationAPIServiceTest {
 		when(apiResponse.isSuccess()).thenReturn(false);
 
 		assertThrows(ContractNegotiationAPIException.class, 
-				()-> service.handleContractNegotiationAccepted(contractNegotaitionId));	}
+				()-> service.handleContractNegotiationAccepted(contractNegotaitionId));	
+		}
+	
+	@Test
+	@DisplayName("Provider accepts contract negotiation")
+	public void handleCNApproved() {
+		String contractNegotaitionId = UUID.randomUUID().toString(); 
+		when(contractNegotiationRepository.findById(contractNegotaitionId)).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_REQUESTED));
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
+		when(properties.providerCallbackAddress()).thenReturn(ModelUtil.CALLBACK_ADDRESS);
+		when(properties.getAssignee()).thenReturn(ModelUtil.ASSIGNEE);
+		when(offerRepository.findById(any(String.class))).thenReturn(Optional.of(ModelUtil.OFFER));
+		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
+		when(apiResponse.isSuccess()).thenReturn(true);
+		
+		service.handleContractNegotiationAgreed(contractNegotaitionId);
+		
+		verify(contractNegotiationRepository).save(argumentCaptor.capture());
+		assertEquals(ContractNegotiationState.AGREED, argumentCaptor.getValue().getState());
+	}
+	
+	@Test
+	@DisplayName("Provider accepts contract negotiation - invalid initial state")
+	public void handleCNApproved_invalid_state() {
+		String contractNegotaitionId = UUID.randomUUID().toString(); 
+		when(contractNegotiationRepository.findById(contractNegotaitionId)).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_AGREED));
+		
+		assertThrows(ContractNegotiationInvalidStateException.class, 
+				() -> service.handleContractNegotiationAgreed(contractNegotaitionId));
+	}
+	
+	@Test
+	@DisplayName("Provider accepts contract negotiation - offer does not exists")
+	public void handleCNApproved_no_offer() {
+		String contractNegotaitionId = UUID.randomUUID().toString(); 
+		when(contractNegotiationRepository.findById(contractNegotaitionId)).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_REQUESTED));
+		when(offerRepository.findById(any(String.class))).thenReturn(Optional.empty());
+		
+		assertThrows(OfferNotFoundException.class, 
+				() -> service.handleContractNegotiationAgreed(contractNegotaitionId));
+	}
+	
+	@Test
+	@DisplayName("Provider accepts contract negotiation - error while contacting consumer")
+	public void handleCNApproved_error_consumer() {
+		String contractNegotaitionId = UUID.randomUUID().toString(); 
+		when(contractNegotiationRepository.findById(contractNegotaitionId)).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_REQUESTED));
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
+		when(properties.providerCallbackAddress()).thenReturn(ModelUtil.CALLBACK_ADDRESS);
+		when(properties.getAssignee()).thenReturn(ModelUtil.ASSIGNEE);
+		when(offerRepository.findById(any(String.class))).thenReturn(Optional.of(ModelUtil.OFFER));
+		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
+		when(apiResponse.isSuccess()).thenReturn(false);
+		
+		assertThrows(ContractNegotiationAPIException.class, 
+				() -> service.handleContractNegotiationAgreed(contractNegotaitionId));
+		
+		verify(contractNegotiationRepository, times(0)).save(argumentCaptor.capture());
+		verify(agreementRepository, times(0)).save(argCaptorAgreement.capture());
+	}
+
 }
